requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiohttp-swagger==1.0.15
aiohttp-swagger3== 0.6.0
async-lru==1.0.2
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
dill==0.3.4
distro==1.5.0
Flask==2.0.1
Flask-Cors==1.10.3
genie==21.6
genie.libs.clean==21.6
genie.libs.conf==21.6
genie.libs.filetransferutils==21.6
genie.libs.health==21.6
genie.libs.ops==21.6
genie.libs.parser==21.6
genie.libs.sdk==21.6
idna==2.10
itsdangerous==2.0.1
Jinja2==2.11.2
jsonpickle==2.0.0
junit-xml==1.9
MarkupSafe==1.1.1
multidict==5.1.0
netaddr==0.8.0
pathspec==0.8.1
prettytable==2.1.0
psutil==5.8.0
pyats==21.6
pyats.aereport==21.6
pyats.aetest==21.6
pyats.async==21.6
pyats.connections==21.6
pyats.datastructures==21.6
pyats.easypy==21.6
pyats.kleenex==21.6
pyats.log==21.6
pyats.reporter==21.6
pyats.results==21.6
pyats.tcl==21.6
pyats.topology==21.6
pyats.utils==21.6
pycparser==2.20
pyftpdlib==1.5.6
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
six==1.16.0
tftpy==0.8.0
tqdm==4.61.1
typing-extensions==3.10.0.0
unicon==21.6
unicon.plugins==21.6
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
xmltodict==0.12.0
yamllint==1.26.1
yarl==1.6.3